refactor: merge plugins and plugins-api
