--- rush-plugins-api/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "rush-plugins-api"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-serde = "1.0.160"
-serde_derive = "1.0.160"
