+++ rush-plugins/Cargo.toml
@@ -5,15 +5,26 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[features]
+host = [
+    "dep:extism",
+    "dep:snafu",
+    "dep:serde_json",
+    "dep:anyhow",
+    "dep:lazy_static",
+    "dep:oneshot",
+]
+default = ["host"]
+
 [dependencies]
-extism = "0.3.0"
-serde = "1.0.160"
-serde_json = "1.0.96"
-snafu = "0.7.4"
-rush-plugins-api = { path = "../rush-plugins-api" }
-anyhow.workspace = true
-lazy_static = "1.4.0"
-oneshot = "0.1.5"
+extism = { version = "0.3.0", optional = true }
+serde = "1.0"
+serde_json = { version = "1.0", optional = true }
+snafu = { version = "0.7.4", optional = true }
+anyhow = { workspace = true, optional = true }
+lazy_static = { version = "1.4.0", optional = true }
+oneshot = { version = "0.1.5", optional = true }
+serde_derive = "1.0"
 
 [dev-dependencies]
 ctor = "0.2.0"
