+++ rush/src/main.rs
@@ -5,9 +5,9 @@ use anyhow::Result;
 use is_executable::is_executable;
 use rush_eval::dispatcher::Dispatcher;
 use rush_eval::errors::DispatchError;
+use rush_plugins::api::InitHookParams;
 use rush_plugins::plugin::HostBindings;
 use rush_plugins::{CurrentPlugin, PluginHost, HOST_BINDINGS};
-use rush_plugins_api::InitHookParams;
 use rush_state::console::{restore_terminal, Console};
 use rush_state::shell::Shell;
 use rush_state::showln;
